requirements.txt
pillow
colorama
pyinstaller